requirements.txt
# PDF Multimodal RAG - Requirements
# Phase 1 & 2

# PDF Parsing
docling>=2.15.0
docling-core>=2.21.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing & Chunking
nltk>=3.8.0
sentence-transformers>=2.2.0  # For semantic chunking similarity

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tqdm>=4.65.0
rich>=13.0.0  # Better console output

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
ipykernel>=6.25.0  # For notebooks
jupyter>=1.0.0

# Vector DB
chromadb>=0.5.0

# AWS
boto3>=1.34.0
botocore>=1.34.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0
